Extract common classic matchers logic
